Fix memory bug causing segmentation fault when reallocating memory
When csmVector is used with std::string, as it is in the Linux
sample code, a segmentation fault occurs. This is due to csmVector
using memcpy when reallocating memory.

A proper way to handle memory would be to explicitly construct
new objects and then destroying the old ones, one by one,
instead of using memcpy().

It is understood that this approach may have performance penalties,
and it is also not well tested for cases where PushBack is called
with callPlacementNew set to false. This commit is provided to
the upstream maintainers to let them decide which approach to fix
the segfault is best for them.